move test to runner suite
